refactor: responseMessages files -> components
