+++ components/movies/messages/moviesResponseMessages.js
@@ -1,0 +1,9 @@
+module.exports = {
+  movieIsAlreadyExistMessage: 'Этот фильм уже сохранён у пользователя.',
+  movieNotFoundMessage: 'Фильма с таким id в базе нет.',
+  movieBelongsToAnotherUserMessage: 'Руки прочь от чужого фильма!',
+  movieDeletedSuccessfullyMessage: 'Фильм удалён.',
+  invalidThumbnailMessage: 'Некорректная ссылка на миниатюру постера к фильму.',
+  invalidImageMessage: 'Некорректная ссылка на постер к фильму.',
+  invalidTrailerMessage:'Некорректная ссылка на трейлер к фильму.',
+};
