+++ components/users/controllers/getUser.js
@@ -1,6 +1,6 @@
 const User = require('../models/user');
 const NotFoundError = require('../../../utils/errors/404');
-const { userNotFoundMessage } = require('../../../utils/responseMessages');
+const { userNotFoundMessage } = require('../messages/usersResponseMessages');
 
 module.exports = async (req, res) => {
   const user = await User.findById(req.user._id).select('-__v');
