+++ components/users/controllers/logout.js
@@ -1,4 +1,4 @@
-const { logoutSuccessfulMessage } = require('../../../utils/responseMessages');
+const { logoutSuccessfulMessage } = require('../messages/usersResponseMessages');
 
 module.exports = (req, res) => {
   res.status(200).clearCookie('jwt').send({ message: logoutSuccessfulMessage });
