+++ components/users/messages/usersResponseMessages.js
@@ -1,0 +1,11 @@
+module.exports = {
+  unauthorizedErrorMessage: 'Ошибка авторизации.',
+  userNotFoundMessage: 'Пользователя с таким id в базе нет.',
+  userTriedToUpdateAnotherProfileMessage: 'Не трожь чужой профиль!',
+  userAlreadyRegisteredMessage: 'Пользователь с таким email уже зарегистрирован.',
+  loginSuccessfulMessage: 'Успешный логин.',
+  logoutSuccessfulMessage: 'Чао-какао!',
+  invalidEmailMessage: 'Некорректный email',
+  invalidRequestDataMessage: 'Переданы некорректные данные.',
+  wrongEmailOrPasswordMessage: 'Неправильные почта или пароль.',
+};
