+++ utils/responseMessages.js
@@ -2,20 +2,5 @@ module.exports = {
   appStartFailedMessage: 'Ошибка запуска приложения.',
   serverErrorMessage: 'Ошибка сервера.',
   resourceNotFoundMessage: 'Запрашиваемый ресурс не найден.',
-  unauthorizedErrorMessage: 'Ошибка авторизации.',
-  userNotFoundMessage: 'Пользователя с таким id в базе нет.',
-  userTriedToUpdateAnotherProfileMessage: 'Не трожь чужой профиль!',
-  userAlreadyRegisteredMessage: 'Пользователь с таким email уже зарегистрирован.',
-  loginSuccessfulMessage: 'Успешный логин.',
-  logoutSuccessfulMessage: 'Чао-какао!',
-  movieIsAlreadyExistMessage: 'Этот фильм уже сохранён у пользователя.',
-  movieNotFoundMessage: 'Фильма с таким id в базе нет.',
-  movieBelongsToAnotherUserMessage: 'Руки прочь от чужого фильма!',
-  movieDeletedSuccessfullyMessage: 'Фильм удалён.',
-  invalidThumbnailMessage: 'Некорректная ссылка на миниатюру постера к фильму.',
-  invalidImageMessage: 'Некорректная ссылка на постер к фильму.',
-  invalidTrailerMessage:'Некорректная ссылка на трейлер к фильму.',
-  invalidEmailMessage: 'Некорректный email',
-  invalidRequestDataMessage: 'Переданы некорректные данные.',
-  wrongEmailOrPasswordMessage: 'Неправильные почта или пароль.',
+  requestValidationErrorMessage: 'Переданы некорректные данные.',
 };
